requirements.txt

sentence-transformers
google-generativeai
python-dotenv
langchain
pinecone-client
flask
langchain_community
ipykernel
pypdf

-e .